Remove deprecated ReorderWindows optimizer
Iterative optimizer is turned on by default so the new SwapAdjacentWindowsByPartitionsOrder rule is used instead.

PruneIdentityProjections does not exist anymore, so we cannot fall back to the ReorderWindows optimizer which required that rewrite as a prerequisite.
Therefore ReorderWindows rule is not needed any more.

New rule does not provide a feature flag which therefore becomes obsolete as well.